Rename and move util.go
